python/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.13.5
kiwisolver==1.3.2
matplotlib==3.5.1
munch==2.5.0
numpy==1.21.5
packaging==21.3
pathtools==0.1.2
Pillow==8.4.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
segmentation-models-pytorch==0.2.1
sentry-sdk==1.5.6
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
termcolor==1.1.0
timm==0.4.12
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.0
typing_extensions==4.0.1
urllib3==1.26.8
wandb==0.12.11
yaspin==2.1.0
